Clean installation
The minimal installation of Sulu still contains some files we don't
need, because we are using it only as a JSON API.

So some HTML Twig templates can be deleted, since we don't use Sulu to
render HTML. Instead a JSON Twig template for the homepage was added,
which can be considered as an API endpoint.

The webspace configuration is also updated to only use a single
language.